+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.0.1
+
 - [browser] feat: Show dropped event url in `blacklistUrl`/`whitelistUrl` debug mode
 - [browser] feat: Use better event description instead of `event_id` for user-facing logs
 - [core] ref: Create common integrations that are exposed on `@sentry/core` and reexposed through `browser`/`node`